+++ packages/hooks/src/react-query/index.ts
@@ -1,0 +1,2 @@
+export * from "./useRegister";
+export * from "./react-query";
