+++ packages/hooks/src/react-query/react-query.tsx
@@ -9,22 +9,10 @@ import {
 	QueryClientProvider,
 	useMutation,
 } from "@tanstack/react-query";
-import { hono_client } from "./hono_client";
+import { hono_client } from "../hono_client";
 import type { RegisterFields } from "@repo/zod-types";
-// import { getTodos, postTodo } from '../my-api'
-
-// Create a client
-const queryClient = new QueryClient();
 
 export { QueryClientProvider, QueryClient };
-
-// export function ReactQueryClientProvider({ children }: { children: React.ReactNode }) {
-//   return (
-//     <QueryClientProvider client={queryClient}>
-//       {children}
-//     </QueryClientProvider>
-//   )
-// }
 
 function makeQueryClient() {
 	return new QueryClient({
@@ -68,30 +56,4 @@ export function QueryProvider({
 	return (
 		<QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
 	);
-}
-
-export function useRegister() {
-	return useMutation({
-		onError: (error) => {
-			console.log(error.message);
-		},
-		onSuccess: (data) => {
-			console.log(data);
-		},
-
-		mutationFn: (data: RegisterFields) => {
-			console.log({
-				data,
-			});
-			return hono_client.api.auth.register.$post({
-				form: {
-					email: data.email,
-					password: data.password,
-					name: data.name,
-					phone: data.phone,
-					agreeTerms: false.toString(),
-				},
-			});
-		},
-	});
 }
