+++ packages/ui/package.json
@@ -1,97 +1,100 @@
 {
-	"name": "@repo/ui",
-	"version": "0.0.0",
-	"private": true,
-	"scripts": {
-		"ui:add": "npx shadcn-ui@latest add",
-		"lint": "eslint .",
-		"storybook": "storybook dev -p 6006",
-		"build-storybook": "storybook build",
-		"chromatic": "npx chromatic --project-token=<CHROMATIC_TOKEN>"
-	},
-	"peerDependencies": {
-		"react": "^18"
-	},
-	"devDependencies": {
-		"@chromatic-com/storybook": "^1.8.0",
-		"@hookform/devtools": "^4.3.1",
-		"@repo/eslint-config": "*",
-		"@repo/typescript-config": "*",
-		"@storybook/addon-essentials": "^8.2.9",
-		"@storybook/addon-interactions": "^8.2.9",
-		"@storybook/addon-links": "^8.2.9",
-		"@storybook/addon-onboarding": "^8.2.9",
-		"@storybook/blocks": "^8.2.9",
-		"@storybook/react": "^8.2.9",
-		"@storybook/react-vite": "^8.2.9",
-		"@storybook/test": "^8.2.9",
-		"@types/node": "^20",
-		"@types/react": "^18",
-		"autoprefixer": "^10.4.19",
-		"chromatic": "^11.10.4",
-		"eslint-plugin-storybook": "^0.8.0",
-		"postcss": "^8.4.38",
-		"postcss-load-config": "^6.0.1",
-		"storybook": "^8.2.9",
-		"tailwindcss": "^3.4.4",
-		"typescript": "^5.4.5",
-		"vite": "^5.4.2"
-	},
-	"dependencies": {
-		"@repo/zod-types": "*",
-		"@repo/hooks": "*",
-		"@hookform/resolvers": "^3.9.0",
-		"@radix-ui/react-accordion": "^1.2.0",
-		"@radix-ui/react-alert-dialog": "^1.1.1",
-		"@radix-ui/react-aspect-ratio": "^1.1.0",
-		"@radix-ui/react-avatar": "^1.1.0",
-		"@radix-ui/react-checkbox": "^1.1.1",
-		"@radix-ui/react-collapsible": "^1.1.0",
-		"@radix-ui/react-context-menu": "^2.2.1",
-		"@radix-ui/react-dialog": "^1.1.1",
-		"@radix-ui/react-dropdown-menu": "^2.1.1",
-		"@radix-ui/react-hover-card": "^1.1.1",
-		"@radix-ui/react-label": "^2.1.0",
-		"@radix-ui/react-menubar": "^1.1.1",
-		"@radix-ui/react-navigation-menu": "^1.2.0",
-		"@radix-ui/react-popover": "^1.1.1",
-		"@radix-ui/react-progress": "^1.1.0",
-		"@radix-ui/react-radio-group": "^1.2.0",
-		"@radix-ui/react-scroll-area": "^1.1.0",
-		"@radix-ui/react-select": "^2.1.1",
-		"@radix-ui/react-separator": "^1.1.0",
-		"@radix-ui/react-slider": "^1.2.0",
-		"@radix-ui/react-slot": "^1.1.0",
-		"@radix-ui/react-switch": "^1.1.0",
-		"@radix-ui/react-tabs": "^1.1.0",
-		"@radix-ui/react-toast": "^1.2.1",
-		"@radix-ui/react-toggle": "^1.1.0",
-		"@radix-ui/react-toggle-group": "^1.1.0",
-		"@radix-ui/react-tooltip": "^1.1.2",
-		"class-variance-authority": "^0.7.0",
-		"clsx": "^2.1.1",
-		"cmdk": "1.0.0",
-		"date-fns": "^3.6.0",
-		"embla-carousel-react": "^8.2.0",
-		"input-otp": "^1.2.4",
-		"lucide-react": "^0.436.0",
-		"next": "^14.2.8",
-		"next-themes": "^0.3.0",
-		"react-day-picker": "8.10.1",
-		"react-hook-form": "^7.53.0",
-		"react-resizable-panels": "^2.1.2",
-		"recharts": "^2.12.7",
-		"sonner": "^1.5.0",
-		"tailwind-merge": "^2.3.0",
-		"tailwindcss-animate": "^1.0.7",
-		"vaul": "^0.9.1",
-		"zod": "^3.23.8"
-	},
-	"exports": {
-		"./globals.css": "./src/globals.css",
-		"./postcss.config": "./postcss.config.mjs",
-		"./tailwind.config": "./tailwind.config.ts",
-		"./lib/*": "./src/lib/*.ts",
-		"./components/*": ["./src/components/*.tsx", "./src/components/*.ts"]
-	}
+  "name": "@repo/ui",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "ui:add": "npx shadcn-ui@latest add",
+    "lint": "eslint .",
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build",
+    "chromatic": "npx chromatic --project-token=<CHROMATIC_TOKEN>"
+  },
+  "peerDependencies": {
+    "react": "^18"
+  },
+  "devDependencies": {
+    "@chromatic-com/storybook": "^1.8.0",
+    "@hookform/devtools": "^4.3.1",
+    "@repo/eslint-config": "*",
+    "@repo/typescript-config": "*",
+    "@storybook/addon-essentials": "^8.2.9",
+    "@storybook/addon-interactions": "^8.2.9",
+    "@storybook/addon-links": "^8.2.9",
+    "@storybook/addon-onboarding": "^8.2.9",
+    "@storybook/blocks": "^8.2.9",
+    "@storybook/react": "^8.2.9",
+    "@storybook/react-vite": "^8.2.9",
+    "@storybook/test": "^8.2.9",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "autoprefixer": "^10.4.19",
+    "chromatic": "^11.10.4",
+    "eslint-plugin-storybook": "^0.8.0",
+    "postcss": "^8.4.38",
+    "postcss-load-config": "^6.0.1",
+    "storybook": "^8.2.9",
+    "tailwindcss": "^3.4.4",
+    "typescript": "^5.4.5",
+    "vite": "^5.4.2"
+  },
+  "dependencies": {
+    "@repo/zod-types": "*",
+    "@repo/hooks": "*",
+    "@hookform/resolvers": "^3.9.0",
+    "@radix-ui/react-accordion": "^1.2.0",
+    "@radix-ui/react-alert-dialog": "^1.1.1",
+    "@radix-ui/react-aspect-ratio": "^1.1.0",
+    "@radix-ui/react-avatar": "^1.1.0",
+    "@radix-ui/react-checkbox": "^1.1.1",
+    "@radix-ui/react-collapsible": "^1.1.0",
+    "@radix-ui/react-context-menu": "^2.2.1",
+    "@radix-ui/react-dialog": "^1.1.1",
+    "@radix-ui/react-dropdown-menu": "^2.1.1",
+    "@radix-ui/react-hover-card": "^1.1.1",
+    "@radix-ui/react-label": "^2.1.0",
+    "@radix-ui/react-menubar": "^1.1.1",
+    "@radix-ui/react-navigation-menu": "^1.2.0",
+    "@radix-ui/react-popover": "^1.1.1",
+    "@radix-ui/react-progress": "^1.1.0",
+    "@radix-ui/react-radio-group": "^1.2.0",
+    "@radix-ui/react-scroll-area": "^1.1.0",
+    "@radix-ui/react-select": "^2.1.1",
+    "@radix-ui/react-separator": "^1.1.0",
+    "@radix-ui/react-slider": "^1.2.0",
+    "@radix-ui/react-slot": "^1.1.0",
+    "@radix-ui/react-switch": "^1.1.0",
+    "@radix-ui/react-tabs": "^1.1.0",
+    "@radix-ui/react-toast": "^1.2.1",
+    "@radix-ui/react-toggle": "^1.1.0",
+    "@radix-ui/react-toggle-group": "^1.1.0",
+    "@radix-ui/react-tooltip": "^1.1.2",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.1",
+    "cmdk": "1.0.0",
+    "date-fns": "^3.6.0",
+    "embla-carousel-react": "^8.2.0",
+    "input-otp": "^1.2.4",
+    "lucide-react": "^0.436.0",
+    "next": "^14.2.8",
+    "next-themes": "^0.3.0",
+    "react-day-picker": "8.10.1",
+    "react-hook-form": "^7.53.0",
+    "react-resizable-panels": "^2.1.2",
+    "recharts": "^2.12.7",
+    "sonner": "^1.5.0",
+    "tailwind-merge": "^2.3.0",
+    "tailwindcss-animate": "^1.0.7",
+    "vaul": "^0.9.1",
+    "zod": "^3.23.8"
+  },
+  "exports": {
+    "./globals.css": "./src/globals.css",
+    "./postcss.config": "./postcss.config.mjs",
+    "./tailwind.config": "./tailwind.config.ts",
+    "./lib/*": "./src/lib/*.ts",
+    "./components/*": [
+      "./src/components/*.tsx",
+      "./src/components/*.ts"
+    ]
+  }
 }
